requirements.txt
fastapi
uvicorn
pydantic
sentence-transformers
Pillow
requests
numpy
python-dotenv
google-genai
sqlalchemy
asyncpg
motor
dnspython
praw
telethon
playwright
opencv-python
scenedetect
pgvector
